fix: sonar bugs

+++ src/main/java/com/osa/openstreetart/util/JwtUtil.java
@@ -19,7 +19,7 @@ public class JwtUtil implements Serializable {
 
 	private static final long serialVersionUID = -2550185165626007488L;
 
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+	public static final int JWT_TOKEN_VALIDITY = 5 * 60 * 60;
 
 	public static final String AUTHORITIES_KEY = "roles";
